[dev] migrate test helpers, fix up syspath in top level conftest so both `python -m pytest` and `pytest` can import modules under `tests/`
